feat: Rename vanilla to core
